test: consolidate github tests to one module
Update imports in tests from core.github to core.api,
remove legacy tests under tests/core/github, and add a new
tests/core/test_api.py covering GitHub API client functionality.
