+++ tests/cli/commands/backup/test_backup_restore.py
@@ -39,7 +39,9 @@ class TestBackupHandler:
 
         # Mock config manager
         mock_config_manager.load_app_config.return_value = sample_app_config
-        mock_config_manager.load_raw_app_config.return_value = sample_app_config
+        mock_config_manager.load_raw_app_config.return_value = (
+            sample_app_config
+        )
 
         args = MagicMock()
         args.app_name = app_name
@@ -88,7 +90,9 @@ class TestBackupHandler:
         )
 
         mock_config_manager.load_app_config.return_value = sample_app_config
-        mock_config_manager.load_raw_app_config.return_value = sample_app_config
+        mock_config_manager.load_raw_app_config.return_value = (
+            sample_app_config
+        )
 
         args = MagicMock()
         args.app_name = app_name
@@ -146,7 +150,9 @@ class TestBackupHandler:
         )
 
         mock_config_manager.load_app_config.return_value = current_config
-        mock_config_manager.load_raw_app_config.return_value = sample_app_config
+        mock_config_manager.load_raw_app_config.return_value = (
+            sample_app_config
+        )
 
         args = MagicMock()
         args.app_name = app_name
