--- tests/core/github/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for infrastructure GitHub module."""
