+++ tests/core/test_conftest.py
@@ -10,7 +10,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from my_unicorn.core.github import Asset, Release
+from my_unicorn.core.api import Asset, Release
 from tests.core.conftest import MockProgressReporter, async_chunk_gen
