+++ tests/core/update/test_context_helpers.py
@@ -11,7 +11,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from my_unicorn.core.github import Asset
+from my_unicorn.core.api import Asset
 from my_unicorn.core.update.context import (
     load_catalog_for_update,
     load_update_config,
