+++ tests/core/update/test_missing_app_update.py
@@ -9,7 +9,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from my_unicorn.core.github import Release
+from my_unicorn.core.api import Release
 from my_unicorn.core.update.info import UpdateInfo
 from my_unicorn.core.update.manager import UpdateManager
