+++ tests/core/update/test_workflows_single_app_advanced.py
@@ -10,7 +10,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from my_unicorn.core.github import Asset, Release
+from my_unicorn.core.api import Asset, Release
 from my_unicorn.core.update.info import UpdateInfo
 from my_unicorn.core.update.workflows import update_single_app
