+++ tests/core/verification/test_detection.py
@@ -15,7 +15,7 @@ from __future__ import annotations
 from typing import Any
 from unittest.mock import MagicMock
 
-from my_unicorn.core.github import Asset
+from my_unicorn.core.api import Asset
 from my_unicorn.core.verification.detection import (
     auto_detect_checksum_files,
     check_digest_availability,
