+++ tests/core/verification/test_execution.py
@@ -16,7 +16,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 
 from my_unicorn.constants import VerificationMethod
-from my_unicorn.core.github import Asset
+from my_unicorn.core.api import Asset
 from my_unicorn.core.verification.context import VerificationContext
 from my_unicorn.core.verification.execution import (
     execute_all_verification_methods,
