+++ tests/core/verification/test_multi_method_verification.py
@@ -15,7 +15,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from my_unicorn.core.github import Asset
+from my_unicorn.core.api import Asset
 from my_unicorn.core.verification.service import VerificationService
 from my_unicorn.exceptions import VerificationError
