+++ tests/integration/test_workflows.py
@@ -20,8 +20,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from my_unicorn.core.api import Asset, Release
 from my_unicorn.core.download import HAS_AIOFILES, DownloadService
-from my_unicorn.core.github import Asset, Release
 from my_unicorn.core.install import InstallHandler
 from my_unicorn.core.post_download import PostDownloadResult
 from my_unicorn.core.protocols.progress import (
